backend/requirements.txt
Django~=3.2.3
djangorestframework~=3.13.1
django-filter~=22.1
django-cors-headers~=3.13.0
django-filter~=22.1
djoser~=2.1.0
psycopg2~=2.9.5
Pillow~=9.3.0
celery~=5.2.7
django-celery-beat~=2.4.0
Redis~=4.4.0